[SUS-46]Refactor/member_api_refactor
