+++ src/member/router/changeRouter.ts
@@ -1,0 +1,49 @@
+import controller from '#controller'
+import express from 'express'
+import NicknameChangeRequest from '../dao/frontend/request/NicknameChangeRequest.js'
+import PasswordChangeRequest from '../dao/frontend/request/PasswordChangeRequest.js'
+import SendFindPasswordEmailRequest from '../dao/frontend/request/SendFindPasswordEmailRequest.js'
+import ChangeService from '../service/ChangeService.js'
+export const changeRouter = express.Router()
+
+/** 닉네임 변경 API */
+changeRouter.patch(
+  '/nickname',
+  controller(
+    null,
+    null,
+    NicknameChangeRequest,
+    null,
+  )(async (req, res): Promise<any> => {
+    await ChangeService.nicknameChange(req, req.body)
+    return res.sendStatus(200)
+  }),
+)
+
+/** 비밀번호 변경 API */
+changeRouter.patch(
+  '/password',
+  controller(
+    null,
+    null,
+    PasswordChangeRequest,
+    null,
+  )(async (req, res) => {
+    await ChangeService.changePassword(req, req.body)
+    return res.sendStatus(200)
+  }),
+)
+
+/** 비밀번호 찾기 인증 이메일 재전송 API */
+changeRouter.get(
+  '/password/email',
+  controller(
+    SendFindPasswordEmailRequest,
+    null,
+    null,
+    null,
+  )(async (req, res) => {
+    await ChangeService.sendPasswordFindEmail(req.query)
+    return res.sendStatus(200)
+  }),
+)
