+++ src/member/service/MemberService.ts
@@ -1,160 +1,12 @@
-import FindPasswordRequest from 'src/member/dao/frontend/request/FindPasswordRequest.js'
-import NicknameChangeRequest from 'src/member/dao/frontend/request/NicknameChangeRequest.js'
-
-import ErrorRegistry from '#error/ErrorRegistry'
-import EmailSender from '#util/email/mailSender/index'
-import Token from '#util/token/index'
 import dotenv from 'dotenv'
-import { Request, response, Response } from 'express'
-import NormalLoginRequest from 'src/member/dao/frontend/request/NormalLoginRequest.js'
-import PasswordChangeRequest from 'src/member/dao/frontend/request/PasswordChangeRequest.js'
-import SignupRequest from 'src/member/dao/frontend/request/SignupRequest.js'
-import SignupVerifyRequest from 'src/member/dao/frontend/request/SignupVerifyRequest.js'
-import MemberRepository from 'src/member/repository/MemberRepository.js'
+import { Response } from 'express'
 
-import jwt from 'jsonwebtoken'
-import SendFindPasswordEmailRequest from 'src/member/dao/frontend/request/SendFindPasswordEmailRequest.js'
-import SendVerifyEmailRequest from 'src/member/dao/frontend/request/SendVerifyEmailRequest.js'
-import SignupResponse from 'src/member/dao/frontend/response/SignupResponse.js'
-import RandomNicknameGenerator from '../../core/util/nickname/nicknameGenerator/index.js'
 dotenv.config()
 
 const signupRedirectUrl = `${process.env.FRONTEND_SERVER_URL}/redirect/signup`
 const loginRedirectUrl = `${process.env.FRONTEND_SERVER_URL}/redirect/login`
 
 export default class MemberService {
-  /** 인증 이메일 전송 서비스 로직 */
-  static async emailSend(signupRequest: SignupRequest) {
-    const { id, password, passwordCheck, email } = signupRequest
-    if (password != passwordCheck) {
-      throw ErrorRegistry.PASSWORD_NOT_EQUAL
-    }
-    await MemberRepository.checkIdAndEmailDuplicate(id, email)
-    await MemberRepository.insertMemberDataAtRedis(id, password, email)
-    const token = Token.generateToken(id, email)
-    EmailSender.sendSignupVerifyEmail(email, token)
-    return new SignupResponse(token)
-  }
-
-  /** 회원가입 인증 서비스 로직 */
-  static async verifySignup(signupVerifyRequest: SignupVerifyRequest) {
-    const { token } = signupVerifyRequest // 쿼리에 포함된 토큰
-    if (!process.env.JWT_SIGNATURE_KEY)
-      throw ErrorRegistry.INTERNAL_SERVER_ERROR
-    const data: any = jwt.verify(token, process.env.JWT_SIGNATURE_KEY)
-    const nickname = await RandomNicknameGenerator.generateNickname() // 랜덤 생성기 자리
-    await MemberRepository.checkVerifyEmailDataFromRedis(data.email)
-    const memberHashData: any = await MemberRepository.getHashDataFromRedis(
-      data.email,
-    )
-    await MemberRepository.insertNormalMemberData(
-      memberHashData.id,
-      memberHashData.password,
-      data.email,
-      nickname,
-    )
-    const signupToken = Token.generateToken(memberHashData.id, data.email)
-    Token.generateCookie('signup', signupToken, response)
-    return signupRedirectUrl
-  }
-  /** 회원가입 인증 이메일 재전송 서비스 로직 */
-  static async sendVerifyEmail(sendVerifyEmailRequest: SendVerifyEmailRequest) {
-    const { id, email } = sendVerifyEmailRequest
-    await MemberRepository.increaseVerifyEmailDataFromRedis(email)
-    const token = Token.generateVerifyToken(id, email)
-    EmailSender.sendSignupVerifyEmail(email, token)
-  }
-
-  /** 닉네임 변경 서비스 로직 */
-  static async nicknameChange(
-    req: Request,
-    nicknameChangeRequest: NicknameChangeRequest,
-  ) {
-    const data = Token.getDataFromToken(req)
-    const { nickname } = nicknameChangeRequest
-    await MemberRepository.changeNickname(nickname, data.userId)
-  }
-
-  /** 비밀번호 검색 서비스 로직 */
-  static async findPassword(findPasswordRequest: FindPasswordRequest) {
-    const { id, email } = findPasswordRequest
-    const checkResult = await MemberRepository.checkMemberPasswordFromDb(
-      id,
-      email,
-    )
-    if (checkResult == 0) {
-      throw ErrorRegistry.CAN_NOT_FIND_USER
-    }
-    const token = Token.generateVerifyToken(id, email)
-    EmailSender.sendFindPasswordVerifyEmail(email, token)
-  }
-
-  /** 비밀번호 변경 서비스 로직 */
-  static async changePassword(
-    req: Request,
-    passwordChangeRequest: PasswordChangeRequest,
-  ) {
-    const { password } = passwordChangeRequest
-    try {
-      const data = Token.getDataFromToken(req)
-      await MemberRepository.updateMemberPassword(password, data.userId)
-      return
-    } catch (e) {
-      throw ErrorRegistry.PASSWORD_CHANGE_FAILED
-    }
-  }
-
-  /** 비밀번호 찾기 인증 이메일 재전송 서비스 로직 */
-  static async sendPasswordFindEmail(
-    sendFindPasswordEmailRequest: SendFindPasswordEmailRequest,
-  ) {
-    const { id, email } = sendFindPasswordEmailRequest
-    const token = Token.generateVerifyToken(id, email)
-    await MemberRepository.checkFindPasswordEmailDataFromRedis(email, token)
-    EmailSender.sendFindPasswordVerifyEmail(email, token)
-  }
-
-  /** 일반 로그인 서비스 로직*/
-  static async attemptNormalLogin(
-    normalLoginRequest: NormalLoginRequest,
-    res: Response,
-  ) {
-    const { id, password } = normalLoginRequest
-    const memberData: any = await MemberRepository.getNormalLoginData(
-      id,
-      password,
-    )
-    if (!memberData) {
-      throw ErrorRegistry.CAN_NOT_FIND_USER
-    }
-    const loginToken: string = Token.generateLoginToken(
-      memberData.userId,
-      memberData.email,
-      memberData.role,
-    )
-    Token.generateCookie('loginToken', loginToken, res)
-    return loginRedirectUrl
-  }
-
-  /** 로그인 상태 체크 */
-  static async checkLogin(req: Request) {
-    if (!req.cookies) {
-      throw ErrorRegistry.LOGIN_REQUIRED
-    }
-    if (!req.cookies.loginToken) {
-      throw ErrorRegistry.LOGIN_REQUIRED
-    }
-    if (!process.env.JWT_SIGNATURE_KEY) {
-      throw ErrorRegistry.INTERNAL_SERVER_ERROR
-    }
-    try {
-      jwt.verify(req.cookies.loginToken, process.env.JWT_SIGNATURE_KEY)
-      return
-    } catch (e) {
-      throw ErrorRegistry.LOGIN_REQUIRED
-    }
-  }
-
   /** 로그아웃 */
   static async logout(res: Response) {
     res.clearCookie('loginToken')
